Added hybrid output layer.
Added hybrid output layer

Worked on transfer-learning paper draft

At transfer learning draft V7

Added 'study_date'

Many changes

+++ mymi/processing/dataset/nifti.py
@@ -99,7 +99,7 @@ def convert_to_training(
     cols = {
         'dataset': str,
         'sample-id': int,
-        'group-id': int,
+        'group-id': float,
         'origin-dataset': str,
         'origin-patient-id': str,
         'region': str,
